Refactor tests
Took 6 minutes

Took 10 minutes
