+++ tests/Helper/ReflectionHelper.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Loper\Tests\Helper;
+
+use ReflectionClass;
+
+class ReflectionHelper
+{
+
+    /**
+     * @throws \ReflectionException
+     */
+    public static function setProperties(ReflectionClass $reflection, object $object, array $properties): void
+    {
+        foreach ($properties as $name => $value) {
+            $reflection_property = $reflection->getProperty($name);
+            $reflection_property->setValue($object, $value);
+        }
+    }
+}
